Refs #1921. Fixes the original reported problem with exiting from MythGallery.
The problem fixed here has to do with a check in the iconview exit code that assumed your cursor would either be on the Gallery dir or the first removable device when you tried to exit the device view. But if there is more than one removable device, this is not always the case.

I'm not closing the ticket because there appears to be another less common problem exiting from MythGallery due to some other bug.



git-svn-id: http://svn.mythtv.org/trunk@10236 4b91aca1-2a1f-0410-a73a-a78f2df169e8