This is the UI side of the feature to allow you to use different re-encoding profiles.
Its useful if you want to squeeze more files on a DVD or if you aren't
happy with the default re-encoding profile you can use a different one.

The new file size indicated is only an estimate because there are a lot
of variables that its not easy to take into account when calculating what
the new size will be after it is re-encoded.

git-svn-id: http://svn.mythtv.org/trunk@10365 4b91aca1-2a1f-0410-a73a-a78f2df169e8