Fixes #2013. Allow user to enter multiple leading zeros in "dumb" channel change.
When optimizing the regular expressions for the slow regular expression matcher on Mac OS X, we made the leading zero regular expression non-greedy. This meant it only eliminated the first of the zeros entered each time GetQueuedChanNum() was called. When smart channel changing is disabled GetQueuedChanNum() was not called on each character entered so only some of the leading zeros were removed.

This implements the original regular expression manually.



git-svn-id: http://svn.mythtv.org/trunk@10482 4b91aca1-2a1f-0410-a73a-a78f2df169e8