Fixes #2044. Start getting MythTV ready for in kernel ivtv driver in Linux 2.6.18 with patch from Janne.
Unless ivtv releases a fixed driver for 2.6.18, the upcoming Linux 2.6.18 release should not be used by those wishing to have fully functional PVR-250 or PVR-350 support, the ivtv driver has not yet been completely ported to the new interface. In particular aspect ratio info, captions and teletext will not be recorded when using the in kernel 2.6.18 ivtv driver.



git-svn-id: http://svn.mythtv.org/trunk@10589 4b91aca1-2a1f-0410-a73a-a78f2df169e8