Fix a problem with MythWelcome where it wasn't updating its status messages after returning to it after exiting the frontend.
It was caused by the way the new MythSocket continues to process incoming 
messages and dispatch events even when the main qt loop is not running.
Previously these messages would be queued up and only processed after the 
frontend had exited and the main loop in MythWelcome was running again.



git-svn-id: http://svn.mythtv.org/trunk@10637 4b91aca1-2a1f-0410-a73a-a78f2df169e8