Fixes #2177. Hides channels hidden in EPG when channel flipping.
I'm not sure if this really is better. I read the arguements from last year and restored the behaviour that was implemented then. However, if you don't like it there is a boolean "skip_non_visible" in channelutil.cpp, if you flip that these channels will only be hidden in the EPG and not in channel flipping.



git-svn-id: http://svn.mythtv.org/trunk@10773 4b91aca1-2a1f-0410-a73a-a78f2df169e8