Refs #1828. Adds back the "UseOutputPictureControls" configuration option.
In order to use the "F" playback picture controls you must enable this option in the Playback Settings. By default it is disabled because the default for the hue value varies between cards and we only know correct defaults for ATI, nVidia, VIA, and S3 cards, for other cards you may need to fix the hue after enabling this option.



git-svn-id: http://svn.mythtv.org/trunk@10789 4b91aca1-2a1f-0410-a73a-a78f2df169e8