Update the help text for mpeg4optionidct and mpeg4optionime options. The real problem is that the code is not processing the data properly and can cause damaged video regardless of CPU usage. These damaged artifacts can be seen at low resolution when the CPU is not even close to being pegged (in fact, the high CPU may be a symptom of the problems while mis-handling the data).
What the user needs to know is that if they turn these on then see
damaged recordings (for whatever reason) they should turn these off
again as they are almost surely the cause of any new problems.

Because these option flags didn't work in previous versions, I agree
that resetting them in dbcheck is a non-intrusive, good solution to
prevent new complaints with 0.20.

Refs #2061


git-svn-id: http://svn.mythtv.org/trunk@11064 4b91aca1-2a1f-0410-a73a-a78f2df169e8