Refs #2320
Adds a new 'watched' flag for recordings. It can be manually set or optionally done automatically.

Optionally allow autoexpire to use this new flag to new watched recordings in preference to unwatched ones.

Optionally prevent autoexpired watched recordings from being re-recorded.

Defaults of the new settings maintain current behaviour.

If you are using a custom theme you will need to upgrade it to work with the new icon on the Watched Recordings page. Custom OSDs may break and cause myth to hang, unless updated to include a fifth textarea in the basedialog container.


git-svn-id: http://svn.mythtv.org/trunk@11138 4b91aca1-2a1f-0410-a73a-a78f2df169e8