Added a "Watch List" feature to the Watch Recordings page.  This creates a list of the recordings that would be most useful to watch in order to catch up on series and shows where I may have fallen behind. Choosing to watch and delete things from the top of this list would help me stay on top of things efficiently without having to examine the full recording list to try to identify these and without overlooking something.
Settings options can be found on the 5th page of TV Settings->Playback.  
There is some verbose info about the filtering and scoring when using
"mythfrontend -v file".

This is turned on by default for now in order to encourage feedback.
However, it may default to off for release if it is believed to be
confusing to many users.


git-svn-id: http://svn.mythtv.org/trunk@11310 4b91aca1-2a1f-0410-a73a-a78f2df169e8