Fixes #2241. Renames the hack for ATI's broken I420 support to be more generic and adds the Intel IEGD driver to the description.
The problem here appears to be a bug in the drivers, wrt the drivers' I420 support.

The YV12 support in these drivers isn't broken, so when this hack is enabled we just make YV12 preferred over I420.

The driver developers probably only tested with mplayer or Xine, which both default to YV12 over I420 when both are available.



git-svn-id: http://svn.mythtv.org/trunk@11372 4b91aca1-2a1f-0410-a73a-a78f2df169e8