Refs #2171. Disable "convert" filter.
This filter is completely broken. It converts to and from 422P incorrectly, and overwrites memory whether you use padded strides or not. 

In order to support any conversion which changes the size of the picture planes we need to change the filter API. Since we don't actually use this filter (all our filters work with 420P/YV12) I've just disabled it instead.



git-svn-id: http://svn.mythtv.org/trunk@11389 4b91aca1-2a1f-0410-a73a-a78f2df169e8