Fixes #2073. This extends support for recovery from a EINVAL or EIO to v4l2 drivers other than the go7007 using modified patch from gnome42 at gmail.
When the bttv v4l2 driver freaks it errors with one of these and you need to reset the buffer state to continue, we already had the code to do this with the go7007 so this just employs this for all v4l2 drivers. This is a safe API call for all v4l2 drivers and is recommended practice anyway. The reset also appears to help when the cx88 freaks out on channel changes (by dropping most of the damaged frames, we still need the retune hack for this driver when we get an EIO).

This does not do a reset on file changes for any drivers other than the go7007, it only does the buffer reset on errors for normal V4L2 devices.

(I removed the cosmetic fixes in the patch to make reading the changeset easier).



git-svn-id: http://svn.mythtv.org/trunk@11416 4b91aca1-2a1f-0410-a73a-a78f2df169e8