Refs #2143. Fix ATSC EIT time offset. (leaving ticket open for possible backport to 0.20-fixes).
The problem was that our offset between Unix time and GPS time was off by five hours combined with a phantom localtime->UTC conversion in QDateTime it made EIT times wrong most places most of the time (annoyingly it was correct for me when I last checked this since my timezone was 5 hours from UTC and 5 + 5 - 5 = 5). We were all compensating for this with a manual timezone setting in mythtv-setup, but obviously this should have been automatic. Now it is...

The DB update sets everyone's timezone compensation to "Auto" for EIT time, and disables the UI for this setting in mythtv-setup. If it turns out to be needed for other reasons I will re-enable it.



git-svn-id: http://svn.mythtv.org/trunk@11430 4b91aca1-2a1f-0410-a73a-a78f2df169e8