Change the 'DisplayGroupDefaultView' combobox popup menu into a group of checkboxes.  You can now have any combination of the following items visible:
  Titles
  Categories
  Recording Groups
  The Watch List
  Search results for search type scheduled recordings

If the user tries turning off all checkboxes and saving, the default of
'Titles only' is automatically selected.

The list of recordings is automatically updated in the background as checkboxes
are turned on and off.  If the user cancels out of the popup the previous
view is restored.  If the user saves the current view, it is also saved to
the database.

The Watch List checkbox on this popup menu also changes the on/off status
on the regular settings page for the Watch List.

The mask value is stored in the settings table as DisplayGroupDefaultViewMask.

On initial loading, a database upgrade will be done to convert the user's
current view settings to the new masked setting type and name.

This patch also removes the 'Default View' and 'Include search rule groups'
setting from the settings page since these are now set and saved from the
popup menu inside Watch Recordings.



git-svn-id: http://svn.mythtv.org/trunk@11440 4b91aca1-2a1f-0410-a73a-a78f2df169e8